fix: clean up create user v3
