+++ internal/command/user_v3_test.go
@@ -83,46 +83,7 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			fields{
 				eventstore: expectEventstore(
 					expectFilter(),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args{
-				ctx: authz.NewMockContext("instanceID", "", ""),
-				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
-				},
-			},
-			res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowPreconditionFailed(nil, "COMMAND-N9QOuN4F7o", "Errors.UserSchema.NotExists"))
-				},
-			},
-		},
-		{
-			"no data, error",
-			fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schema.NewCreatedEvent(
-								context.Background(),
-								&schema.NewAggregate("id1", "instanceID").Aggregate,
-								"type",
-								json.RawMessage(`{
-								"$schema": "urn:zitadel:schema:v1",
-								"type": "object",
-								"properties": {
-									"name": {
-										"type": "string"
-									}
-								}
-							}`),
-								[]domain.AuthenticatorType{domain.AuthenticatorTypeUsername},
-							),
-						),
-					),
+					expectFilter(),
 				),
 				checkPermission: newMockPermissionCheckAllowed(),
 				idGenerator:     mock.ExpectID(t, "id1"),
@@ -130,9 +91,31 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "", ""),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
+				},
+			},
+			res{
+				err: func(err error) bool {
+					return errors.Is(err, zerrors.ThrowPreconditionFailed(nil, "COMMAND-VLDTtxT3If", "Errors.UserSchema.NotExists"))
+				},
+			},
+		},
+		{
+			"no data, error",
+			fields{
+				eventstore: expectEventstore(
+					expectFilter(),
+					filterSchemaExisting(),
+				),
+				checkPermission: newMockPermissionCheckAllowed(),
+				idGenerator:     mock.ExpectID(t, "id1"),
+			},
+			args{
+				ctx: authz.NewMockContext("instanceID", "", ""),
+				user: &CreateSchemaUser{
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 				},
 			},
 			res{
@@ -145,25 +128,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user create, no permission",
 			fields{
 				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schema.NewCreatedEvent(
-								context.Background(),
-								&schema.NewAggregate("id1", "instanceID").Aggregate,
-								"type",
-								json.RawMessage(`{
-								"$schema": "urn:zitadel:schema:v1",
-								"type": "object",
-								"properties": {
-									"name": {
-										"type": "string"
-									}
-								}
-							}`),
-								[]domain.AuthenticatorType{domain.AuthenticatorTypeUsername},
-							),
-						),
-					),
+					expectFilter(),
+					filterSchemaExisting(),
 				),
 				checkPermission: newMockPermissionCheckNotAllowed(),
 				idGenerator:     mock.ExpectID(t, "id1"),
@@ -171,9 +137,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "", ""),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"name": "user"
 					}`),
@@ -189,26 +154,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user created",
 			fields{
 				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schema.NewCreatedEvent(
-								context.Background(),
-								&schema.NewAggregate("id1", "instanceID").Aggregate,
-								"type",
-								json.RawMessage(`{
-								"$schema": "urn:zitadel:schema:v1",
-								"type": "object",
-								"properties": {
-									"name": {
-										"type": "string"
-									}
-								}
-							}`),
-								[]domain.AuthenticatorType{domain.AuthenticatorTypeUsername},
-							),
-						),
-					),
 					expectFilter(),
+					filterSchemaExisting(),
 					expectPush(
 						schemauser.NewCreatedEvent(
 							context.Background(),
@@ -227,9 +174,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "", ""),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"name": "user"
 					}`),
@@ -246,6 +192,7 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user create, no field permission as admin",
 			fields{
 				eventstore: expectEventstore(
+					expectFilter(),
 					expectFilter(
 						eventFromEventPusher(
 							schema.NewCreatedEvent(
@@ -275,9 +222,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "", ""),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"name": "user"
 					}`),
@@ -293,6 +239,7 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user create, no field permission as user",
 			fields{
 				eventstore: expectEventstore(
+					expectFilter(),
 					expectFilter(
 						eventFromEventPusher(
 							schema.NewCreatedEvent(
@@ -321,9 +268,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "org1", "id1"),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"name": "user"
 					}`),
@@ -339,25 +285,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user create, invalid data type",
 			fields{
 				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schema.NewCreatedEvent(
-								context.Background(),
-								&schema.NewAggregate("id1", "instanceID").Aggregate,
-								"type",
-								json.RawMessage(`{
-								"$schema": "urn:zitadel:schema:v1",
-								"type": "object",
-								"properties": {
-									"name": {
-										"type": "string"
-									}
-								}
-							}`),
-								[]domain.AuthenticatorType{domain.AuthenticatorTypeUsername},
-							),
-						),
-					),
+					expectFilter(),
+					filterSchemaExisting(),
 				),
 				checkPermission: newMockPermissionCheckAllowed(),
 				idGenerator:     mock.ExpectID(t, "id1"),
@@ -365,9 +294,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "org1", "user1"),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"name": 1
 					}`),
@@ -383,6 +311,7 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user created, additional property",
 			fields{
 				eventstore: expectEventstore(
+					expectFilter(),
 					expectFilter(
 						eventFromEventPusher(
 							schema.NewCreatedEvent(
@@ -402,7 +331,6 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 							),
 						),
 					),
-					expectFilter(),
 					expectPush(
 						schemauser.NewCreatedEvent(
 							context.Background(),
@@ -421,9 +349,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "", ""),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"additional": "property"
 					}`),
@@ -440,6 +367,7 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user create, invalid data attribute name",
 			fields{
 				eventstore: expectEventstore(
+					expectFilter(),
 					expectFilter(
 						eventFromEventPusher(
 							schema.NewCreatedEvent(
@@ -467,9 +395,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "org1", "user1"),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"invalid": "user"
 					}`),
@@ -485,26 +412,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user created, email return",
 			fields{
 				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schema.NewCreatedEvent(
-								context.Background(),
-								&schema.NewAggregate("id1", "instanceID").Aggregate,
-								"type",
-								json.RawMessage(`{
-								"$schema": "urn:zitadel:schema:v1",
-								"type": "object",
-								"properties": {
-									"name": {
-										"type": "string"
-									}
-								}
-							}`),
-								[]domain.AuthenticatorType{domain.AuthenticatorTypeUsername},
-							),
-						),
-					),
 					expectFilter(),
+					filterSchemaExisting(),
 					expectPush(
 						schemauser.NewCreatedEvent(
 							context.Background(),
@@ -540,9 +449,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "", ""),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"name": "user"
 					}`),
@@ -565,26 +473,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user created, email to verify",
 			fields{
 				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schema.NewCreatedEvent(
-								context.Background(),
-								&schema.NewAggregate("id1", "instanceID").Aggregate,
-								"type",
-								json.RawMessage(`{
-								"$schema": "urn:zitadel:schema:v1",
-								"type": "object",
-								"properties": {
-									"name": {
-										"type": "string"
-									}
-								}
-							}`),
-								[]domain.AuthenticatorType{domain.AuthenticatorTypeUsername},
-							),
-						),
-					),
 					expectFilter(),
+					filterSchemaExisting(),
 					expectPush(
 						schemauser.NewCreatedEvent(
 							context.Background(),
@@ -620,9 +510,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "", ""),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"name": "user"
 					}`),
@@ -643,26 +532,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user created, phone return",
 			fields{
 				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schema.NewCreatedEvent(
-								context.Background(),
-								&schema.NewAggregate("id1", "instanceID").Aggregate,
-								"type",
-								json.RawMessage(`{
-								"$schema": "urn:zitadel:schema:v1",
-								"type": "object",
-								"properties": {
-									"name": {
-										"type": "string"
-									}
-								}
-							}`),
-								[]domain.AuthenticatorType{domain.AuthenticatorTypeUsername},
-							),
-						),
-					),
 					expectFilter(),
+					filterSchemaExisting(),
 					expectPush(
 						schemauser.NewCreatedEvent(
 							context.Background(),
@@ -697,9 +568,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "", ""),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"name": "user"
 					}`),
@@ -721,26 +591,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user created, phone to verify",
 			fields{
 				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schema.NewCreatedEvent(
-								context.Background(),
-								&schema.NewAggregate("id1", "instanceID").Aggregate,
-								"type",
-								json.RawMessage(`{
-								"$schema": "urn:zitadel:schema:v1",
-								"type": "object",
-								"properties": {
-									"name": {
-										"type": "string"
-									}
-								}
-							}`),
-								[]domain.AuthenticatorType{domain.AuthenticatorTypeUsername},
-							),
-						),
-					),
 					expectFilter(),
+					filterSchemaExisting(),
 					expectPush(
 						schemauser.NewCreatedEvent(
 							context.Background(),
@@ -775,9 +627,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "", ""),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"name": "user"
 					}`),
@@ -797,26 +648,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			"user created, full verified",
 			fields{
 				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schema.NewCreatedEvent(
-								context.Background(),
-								&schema.NewAggregate("id1", "instanceID").Aggregate,
-								"type",
-								json.RawMessage(`{
-								"$schema": "urn:zitadel:schema:v1",
-								"type": "object",
-								"properties": {
-									"name": {
-										"type": "string"
-									}
-								}
-							}`),
-								[]domain.AuthenticatorType{domain.AuthenticatorTypeUsername},
-							),
-						),
-					),
 					expectFilter(),
+					filterSchemaExisting(),
 					expectPush(
 						schemauser.NewCreatedEvent(
 							context.Background(),
@@ -849,9 +682,8 @@ func TestCommands_CreateSchemaUser(t *testing.T) {
 			args{
 				ctx: authz.NewMockContext("instanceID", "", ""),
 				user: &CreateSchemaUser{
-					ResourceOwner:  "org1",
-					SchemaID:       "type",
-					schemaRevision: 1,
+					ResourceOwner: "org1",
+					SchemaID:      "type",
 					Data: json.RawMessage(`{
 						"name": "user"
 					}`),
@@ -1093,848 +925,6 @@ func TestCommandSide_DeleteSchemaUser(t *testing.T) {
 				checkPermission: tt.fields.checkPermission,
 			}
 			got, err := r.DeleteSchemaUser(tt.args.ctx, tt.args.orgID, tt.args.userID)
-			if tt.res.err == nil {
-				assert.NoError(t, err)
-			}
-			if tt.res.err != nil && !tt.res.err(err) {
-				t.Errorf("got wrong err: %v ", err)
-			}
-			if tt.res.err == nil {
-				assertObjectDetails(t, tt.res.want, got)
-			}
-		})
-	}
-}
-
-func TestCommandSide_LockSchemaUser(t *testing.T) {
-	type fields struct {
-		eventstore      func(*testing.T) *eventstore.Eventstore
-		checkPermission domain.PermissionCheck
-	}
-	type (
-		args struct {
-			ctx    context.Context
-			orgID  string
-			userID string
-		}
-	)
-	type res struct {
-		want *domain.ObjectDetails
-		err  func(error) bool
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		args   args
-		res    res
-	}{
-		{
-			name: "userid missing, invalid argument error",
-			fields: fields{
-				eventstore:      expectEventstore(),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowInvalidArgument(nil, "COMMAND-Eu8I2VAfjF", "Errors.IDMissing"))
-				},
-			},
-		},
-		{
-			name: "user not existing, not found error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-G4LOrnjY7q", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "user removed, not found error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-						eventFromEventPusher(
-							schemauser.NewDeletedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-G4LOrnjY7q", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "user locked, precondition error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-						eventFromEventPusher(
-							schemauser.NewLockedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-G4LOrnjY7q", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "lock user, ok",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-					),
-					expectPush(
-						schemauser.NewLockedEvent(context.Background(),
-							&schemauser.NewAggregate("user1", "org1").Aggregate,
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				want: &domain.ObjectDetails{
-					ResourceOwner: "org1",
-				},
-			},
-		},
-		{
-			name: "lock user, no permission",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckNotAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowPermissionDenied(nil, "AUTHZ-HKJD33", "Errors.PermissionDenied"))
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			r := &Commands{
-				eventstore:      tt.fields.eventstore(t),
-				checkPermission: tt.fields.checkPermission,
-			}
-			got, err := r.LockSchemaUser(tt.args.ctx, tt.args.orgID, tt.args.userID)
-			if tt.res.err == nil {
-				assert.NoError(t, err)
-			}
-			if tt.res.err != nil && !tt.res.err(err) {
-				t.Errorf("got wrong err: %v ", err)
-			}
-			if tt.res.err == nil {
-				assertObjectDetails(t, tt.res.want, got)
-			}
-		})
-	}
-}
-
-func TestCommandSide_UnlockSchemaUser(t *testing.T) {
-	type fields struct {
-		eventstore      func(*testing.T) *eventstore.Eventstore
-		checkPermission domain.PermissionCheck
-	}
-	type (
-		args struct {
-			ctx    context.Context
-			orgID  string
-			userID string
-		}
-	)
-	type res struct {
-		want *domain.ObjectDetails
-		err  func(error) bool
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		args   args
-		res    res
-	}{
-		{
-			name: "userid missing, invalid argument error",
-			fields: fields{
-				eventstore:      expectEventstore(),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowInvalidArgument(nil, "COMMAND-krXtYscQZh", "Errors.IDMissing"))
-				},
-			},
-		},
-		{
-			name: "user not existing, not found error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-gpBv46Lh9m", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "user removed, not found error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-						eventFromEventPusher(
-							schemauser.NewDeletedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-gpBv46Lh9m", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "user not locked, precondition error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-gpBv46Lh9m", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "unlock user, ok",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-						eventFromEventPusher(
-							schemauser.NewLockedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-							),
-						),
-					),
-					expectPush(
-						schemauser.NewUnlockedEvent(context.Background(),
-							&schemauser.NewAggregate("user1", "org1").Aggregate,
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				want: &domain.ObjectDetails{
-					ResourceOwner: "org1",
-				},
-			},
-		},
-		{
-			name: "unlock user, no permission",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-						eventFromEventPusher(
-							schemauser.NewLockedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckNotAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowPermissionDenied(nil, "AUTHZ-HKJD33", "Errors.PermissionDenied"))
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			r := &Commands{
-				eventstore:      tt.fields.eventstore(t),
-				checkPermission: tt.fields.checkPermission,
-			}
-			got, err := r.UnlockSchemaUser(tt.args.ctx, tt.args.orgID, tt.args.userID)
-			if tt.res.err == nil {
-				assert.NoError(t, err)
-			}
-			if tt.res.err != nil && !tt.res.err(err) {
-				t.Errorf("got wrong err: %v ", err)
-			}
-			if tt.res.err == nil {
-				assertObjectDetails(t, tt.res.want, got)
-			}
-		})
-	}
-}
-
-func TestCommandSide_DeactivateSchemaUser(t *testing.T) {
-	type fields struct {
-		eventstore      func(*testing.T) *eventstore.Eventstore
-		checkPermission domain.PermissionCheck
-	}
-	type (
-		args struct {
-			ctx    context.Context
-			orgID  string
-			userID string
-		}
-	)
-	type res struct {
-		want *domain.ObjectDetails
-		err  func(error) bool
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		args   args
-		res    res
-	}{
-		{
-			name: "userid missing, invalid argument error",
-			fields: fields{
-				eventstore:      expectEventstore(),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowInvalidArgument(nil, "COMMAND-pjJhge86ZV", "Errors.IDMissing"))
-				},
-			},
-		},
-		{
-			name: "user not existing, not found error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-Ob6lR5iFTe", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "user removed, not found error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-						eventFromEventPusher(
-							schemauser.NewDeletedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-Ob6lR5iFTe", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "user not active, precondition error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-						eventFromEventPusher(
-							schemauser.NewDeactivatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-Ob6lR5iFTe", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "deactivate user, ok",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-					),
-					expectPush(
-						schemauser.NewDeactivatedEvent(context.Background(),
-							&schemauser.NewAggregate("user1", "org1").Aggregate,
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				want: &domain.ObjectDetails{
-					ResourceOwner: "org1",
-				},
-			},
-		},
-		{
-			name: "deactivate user, no permission",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckNotAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowPermissionDenied(nil, "AUTHZ-HKJD33", "Errors.PermissionDenied"))
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			r := &Commands{
-				eventstore:      tt.fields.eventstore(t),
-				checkPermission: tt.fields.checkPermission,
-			}
-			got, err := r.DeactivateSchemaUser(tt.args.ctx, tt.args.orgID, tt.args.userID)
-			if tt.res.err == nil {
-				assert.NoError(t, err)
-			}
-			if tt.res.err != nil && !tt.res.err(err) {
-				t.Errorf("got wrong err: %v ", err)
-			}
-			if tt.res.err == nil {
-				assertObjectDetails(t, tt.res.want, got)
-			}
-		})
-	}
-}
-
-func TestCommandSide_ReactivateSchemaUser(t *testing.T) {
-	type fields struct {
-		eventstore      func(*testing.T) *eventstore.Eventstore
-		checkPermission domain.PermissionCheck
-	}
-	type (
-		args struct {
-			ctx    context.Context
-			orgID  string
-			userID string
-		}
-	)
-	type res struct {
-		want *domain.ObjectDetails
-		err  func(error) bool
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		args   args
-		res    res
-	}{
-		{
-			name: "userid missing, invalid argument error",
-			fields: fields{
-				eventstore:      expectEventstore(),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowInvalidArgument(nil, "COMMAND-17XupGvxBJ", "Errors.IDMissing"))
-				},
-			},
-		},
-		{
-			name: "user not existing, not found error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-rQjbBr4J3j", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "user removed, not found error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-						eventFromEventPusher(
-							schemauser.NewDeletedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-rQjbBr4J3j", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "user not inactive, precondition error",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowNotFound(nil, "COMMAND-rQjbBr4J3j", "Errors.User.NotFound"))
-				},
-			},
-		},
-		{
-			name: "activate user, ok",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-						eventFromEventPusher(
-							schemauser.NewDeactivatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-							),
-						),
-					),
-					expectPush(
-						schemauser.NewActivatedEvent(context.Background(),
-							&schemauser.NewAggregate("user1", "org1").Aggregate,
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				want: &domain.ObjectDetails{
-					ResourceOwner: "org1",
-				},
-			},
-		},
-		{
-			name: "activate user, no permission",
-			fields: fields{
-				eventstore: expectEventstore(
-					expectFilter(
-						eventFromEventPusher(
-							schemauser.NewCreatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-								"schema",
-								1,
-								json.RawMessage(`{
-						"name": "user"
-					}`),
-							),
-						),
-						eventFromEventPusher(
-							schemauser.NewDeactivatedEvent(context.Background(),
-								&schemauser.NewAggregate("user1", "org1").Aggregate,
-							),
-						),
-					),
-				),
-				checkPermission: newMockPermissionCheckNotAllowed(),
-			},
-			args: args{
-				ctx:    context.Background(),
-				userID: "user1",
-			},
-			res: res{
-				err: func(err error) bool {
-					return errors.Is(err, zerrors.ThrowPermissionDenied(nil, "AUTHZ-HKJD33", "Errors.PermissionDenied"))
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			r := &Commands{
-				eventstore:      tt.fields.eventstore(t),
-				checkPermission: tt.fields.checkPermission,
-			}
-			got, err := r.ActivateSchemaUser(tt.args.ctx, tt.args.orgID, tt.args.userID)
 			if tt.res.err == nil {
 				assert.NoError(t, err)
 			}
